feat: refactor NodePalette component and implement useNodePalette composable for better state management
