+++ frontend/index.html
@@ -9,7 +9,7 @@
     <!-- Add Font Awesome CSS -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" />
   </head>
-  <body style="margin: 0; padding: 0; font-family: Arial, sans-serif;">
+  <body class="m-0 p-0 font-sans">
     <div id="app"></div>
     <script type="module" src="/src/main.ts"></script>
   </body>
